Module Interpreter__Compile_program

This module takes the typed AST of a bolt program and converts it to a list of instructions to be executed by the interpreter, as well as setting the initial stack and heap of the interpreter.

val compile_program : Typing_core_lang.Typed_ast.program -> (Interpreter.Runtime_env.bytecode * Interpreter.Runtime_env.stack * Interpreter.Runtime_env.heap) Core.Or_error.t