Module Interpreter__Execute_program

The interpreter takes as input the typed AST of the program and first compiles it to an intermediate bytecode representation and an initial state.

It then executes these instructions and optionally prints the state at each step of the execution, before finally returning the final value if the computation succeeds, or a runtime error

val execute_program : Typing.Typed_ast.program -> print_execution:Stdlib.Format.formatter option -> Interpreter.Runtime_env.value Core.Or_error.t

val print_result : Stdlib.Format.formatter -> Interpreter.Runtime_env.value -> unit